src/main/java/yandex/order/CreateOrder.java
package yandex.order;

import io.qameta.allure.Step;
import yandex.ScooterApiClient;

public class CreateOrder extends ScooterApiClient {
    private final String ROOT = "/api/v1/orders";
    private final String firstName;
    private final String lastName;
    private final String address;
    private final String metroStation;
    private final String phone;
    private final int rentTime;
    private final String deliveryDate;
    private final String comment;
    private final String[] color;

    public CreateOrder(String firstName, String lastName, String address, String metroStation, String phone, int rentTime, String deliveryDate, String comment, String[] color) {
        this.firstName = firstName;
        this.lastName = lastName;
        this.address = address;
        this.metroStation = metroStation;
        this.phone = phone;
        this.rentTime = rentTime;
        this.deliveryDate = deliveryDate;
        this.comment = comment;
        this.color = color;
    }

    @Step("Создание заказа")
    public int createOrder() {
        return getSpec()
                .body(this)
                .when()
                .post(ROOT)
                .then().log().all()
                .assertThat()
                .statusCode(201)
                .extract().path("track");
    }
}
